requirements/base.txt
django==5.1.4
pymysql==1.1.1
cryptography==44.0.0
requests==2.32.3
slack-sdk==3.34.0
pytest==8.3.4
pytest-django==4.9.0
pytest-mock==3.14.0
django-model-utils==5.0.0
black[d]==24.10.0
djangorestframework==3.15.2
python-dotenv==1.0.1
aiobotocore[boto3]==2.16.0
